Add files via upload
This was the first project I did during my freshmen year of college where the goal was to Write a C++ program that computes a Total of 5 numbers {12098, 30982, 10983, 8983, 20392} and The percentage each number represents of the total number. I am just going off of the description at hand for the assignment.